Added ability to choose directory from the tray menu.
Customizeable directories is partially done. I can't figure out how to
show the drop down menu, but it is accessible from the tray menu (when
minimized). It still defaults to "C:/temp/TestSyncDir" on initial run,
but it can be changed and it will sync correctly. I need to create some
kind of file that can store the user's preferred directory so we can use
that on startup.